# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/vehicles": {
      "get": {
        "operationId": "VehiclesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar mis vehículos con sus vencimientos",
        "tags": [
          "Vehículos"
        ]
      },
      "post": {
        "description": "El límite por cuenta se controla con MAX_VEHICLES_PER_USER (Fase 1: 1).",
        "operationId": "VehiclesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Alcanzaste el límite de vehículos del plan"
          },
          "409": {
            "description": "Ya registraste esa patente"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar un vehículo",
        "tags": [
          "Vehículos"
        ]
      }
    },
    "/api/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no es tuyo"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver un vehículo",
        "tags": [
          "Vehículos"
        ]
      },
      "patch": {
        "description": "Si se envía \"expirations\", reemplaza por completo las fechas guardadas.",
        "operationId": "VehiclesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un vehículo",
        "tags": [
          "Vehículos"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un vehículo con sus documentos y fechas",
        "tags": [
          "Vehículos"
        ]
      }
    },
...
    "/api/reminders/preview": {
      "get": {
        "description": "No envía nada. Sirve para revisar el contenido de los recordatorios.",
        "operationId": "RemindersController_preview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver qué avisos tocarían hoy para mis vehículos",
        "tags": [
          "Recordatorios"
        ]
      }
    },
    "/api/reminders/run": {
      "post": {
        "description": "Pensado para un cron externo (Railway Scheduled Job). La operación es idempotente: repetirla en el mismo día no duplica avisos.",
        "operationId": "RemindersController_run",
        "parameters": [
          {
            "name": "x-cron-secret",
            "required": true,
            "in": "header",
            "description": "Debe coincidir con CRON_SECRET. Sin esa variable el endpoint queda deshabilitado.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumen de lo enviado"
          },
          "403": {
            "description": "Secreto ausente o incorrecto"
          }
        },
        "summary": "Disparar el envío de recordatorios",
        "tags": [
          "Recordatorios"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mi Revisión Técnica — API",
    "description": "Contrato generado desde el código. Ver README.md.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ExpirationResponse": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "REVISION_TECNICA",
              "SOAP",
              "PERMISO_CIRCULACION",
              "OTRO"
            ]
          },
          "dueDate": {
            "type": "string",
            "example": "2026-09-30"
          },
          "daysRemaining": {
            "type": "number",
            "example": 28,
            "description": "Días que faltan. Negativo si ya venció."
          },
          "status": {
            "type": "string",
            "enum": [
              "vigente",
              "por_vencer",
              "vencido"
            ]
          }
        },
        "required": [
          "kind",
          "dueDate",
          "daysRemaining",
          "status"
        ]
      },
      "VehicleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "plate": {
            "type": "string",
            "example": "ABCD12"
          },
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "object",
            "nullable": true
          },
          "expirations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpirationResponse"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plate",
          "brand",
          "model",
          "expirations",
          "createdAt"
        ]
      },
      "ExpirationDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "REVISION_TECNICA",
              "SOAP",
              "PERMISO_CIRCULACION",
              "OTRO"
            ],
            "example": "REVISION_TECNICA"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-09-30",
            "description": "Fecha de vencimiento (YYYY-MM-DD)."
          }
        },
        "required": [
          "kind",
          "dueDate"
        ]
      },
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "ABCD12",
            "description": "Sin guiones ni espacios."
          },
          "brand": {
            "type": "string",
            "example": "Chevrolet"
          },
          "model": {
            "type": "string",
            "example": "Cruze LT"
          },
          "year": {
            "type": "number",
            "example": 2018
          },
          "expirations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpirationDto"
            }
          }
        },
        "required": [
          "plate",
          "brand",
          "model"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "ABCD12",
            "description": "Sin guiones ni espacios."
          },
          "brand": {
            "type": "string",
            "example": "Chevrolet"
          },
          "model": {
            "type": "string",
            "example": "Cruze LT"
          },
          "year": {
            "type": "number",
            "example": 2018
          },
          "expirations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpirationDto"
            }
          }
        }
      },
...
    }
  }
}